backend/requirements.txt
astroid==2.8.5
attrs==21.2.0
autopep8==1.6.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
Django==2.2.16
django-csvimport==2.16
django-cors-headers==3.9.0
django-filter==21.1
django-colorfield==0.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
idna==3.3
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
packaging==21.2
platformdirs==2.4.0
pluggy==0.13.1
py==1.10.0
pycodestyle==2.8.0
PyJWT==2.3.0
pylint==2.11.1
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2021.3
reportlab==3.6.1
requests==2.26.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
wrapt==1.13.3
gunicorn==20.0.4
psycopg2-binary==2.8.6
python-dotenv==0.19.2
drf-extra-fields==3.4.0